refactor: extraction des projets-protocole par serveur (T0 -> T10)
Réarchitecture fine-grained : les paquets wire quittent les projets applicatifs
pour trois projets-protocole purs, et le générateur devient indifférent au layout.

T0 — baseline verte
  Le working tree portait un refactor à moitié appliqué qui cassait trois choses :
  `partial` retiré de 328 structs wire (viole FEN001), renommage ZoneRuntime -> Zone
  réintroduisant la collision namespace/type de F6 (CS0118 x409), et un `using` perdu.
  Triage : sur 762 .cs modifiés, 52 portaient une instruction réellement ajoutée ou
  supprimée. Réparé plutôt que jeté — le tree contenait le fix des ports de handover,
  le durcissement du bind Center et un Lazy<ICenterLink>.

T1-T4 — générateur indifférent au layout
  FieldModel.Offset supprimé (état mutable dans une clé de cache incrémental, jamais lu).
  22 FQN codés en dur rapatriés dans WellKnownNames. Le namespace d'émission vient
  désormais de Compilation.AssemblyName ; les agrégats sont préfixés par serveur.
  IFrameDispatcher et IOpcodeFrameSizeProvider descendus dans Fenrir.Core.Abstractions :
  c'est l'arête qui empêchait des projets Protocol.* purs d'exister.
  Nouveaux diagnostics, tous Error, zéro violation : FEN005 (taille wire explicite),
  FEN006 (un seul serveur de paquets par assembly), FEN007 (taille imbriquée
  irrésolvable — remplace un retour 0 SILENCIEUX qui désynchronisait la session),
  FEN008 (un seul serveur de handlers par assembly).

T5-T7 — bugs actifs
  SessionId collisionnait entre zones : le compteur vivait sur le listener alors que
  GameConnectionHost en arme un par map et que trois dictionnaires process-wide
  indexent dessus. Armement des listeners désormais tout-ou-rien.

T8-T10 — les trois projets-protocole
  Fenrir.Protocol.Center (4 paquets), .Login (27), .Game (246 + ZoneSessionState).
  Chacun ne référence que Fenrir.Core et le générateur en analyzer.
  Les handlers restent dans les projets applicatifs, qui seuls émettent le dispatch.

Vérification, sans aucun test dans le dépôt : 18 compteurs et 7 empreintes du code
généré, insensibles au renommage par construction. Toutes identiques avant/après,
dont celle des 246 corps de sérialiseurs — seule défense contre un réordonnancement
de champs à largeur égale, invisible à tout le reste puisque l'ordre de déclaration
EST l'ordre wire et que le protocole n'a pas de préfixe de longueur.

20 projets. Build --no-incremental vert, publish AOT vert sur les 4 exécutables.

+++ src/3_Infrastructure/Fenrir.Cluster/WorldState/CenterHeroRankStore.cs
@@ -7,19 +7,6 @@ using Fenrir.Data.Abstractions.Progression;
 
 namespace Fenrir.Cluster.WorldState;
 
-/// <summary>
-///     CaeriusNet-backed persistence for the Center hero-rank current-period accrual. Reuses the existing
-///     current-period procedures: <c>usp_HeroRanking_GetByPeriod</c> (PeriodKind 0) for the boot load and
-///     <c>usp_HeroRanking_Upsert</c> (RewardClaimed/Description left NULL) for the wholesale write.
-/// </summary>
-/// <remarks>
-///     Deliberately does NOT call <c>usp_HeroRanking_AddPoints</c>: its UPDATE branch freezes the stored Level
-///     (the shard-side replication of legacy Bug 1). <see cref="HeroRankAuthority" /> owns the authoritative
-///     running total and monotonic level in memory and flushes it wholesale through the idempotent Upsert, which
-///     corrects Bug 1. The current-period row carries no reward-claim/description state (those belong to the
-///     previous period after rollover), so passing them as NULL on the Upsert is intended, per
-///     <see cref="ICenterHeroRankStore" />.
-/// </remarks>
 public sealed record CenterHeroRankStore(ICaeriusNetDbContext Db) : ICenterHeroRankStore
 {
     private const byte CurrentPeriodKind = 0;
@@ -38,7 +25,7 @@ public sealed record CenterHeroRankStore(ICaeriusNetDbContext Db) : ICenterHeroR
         for (var i = 0; i < rows.Count; i++)
         {
             var row = rows[i];
-            accruals[i] = new CenterHeroRankAccrual(row.CharacterId, (byte)(row.TribeId ?? 0), row.Points,
+            accruals[i] = new CenterHeroRankAccrual(row.CharacterId, row.TribeId ?? 0, row.Points,
                 row.Level ?? 0);
         }
 
@@ -48,8 +35,6 @@ public sealed record CenterHeroRankStore(ICaeriusNetDbContext Db) : ICenterHeroR
     public ValueTask UpsertCurrentPeriodAsync(int characterId, int points, byte tribeId, int level,
         CancellationToken ct)
     {
-        // RewardClaimed and Description are intentionally omitted -- both default to NULL in the procedure,
-        // which is the correct current-period state (see the type remarks above).
         var sp = new StoredProcedureParametersBuilder("game", "usp_HeroRanking_Upsert", 0)
             .AddParameter("CharacterId", characterId, SqlDbType.Int)
             .AddParameter("PeriodKind", CurrentPeriodKind, SqlDbType.TinyInt)
